- final new Linux compiled binaries for 2008b
git-svn-id: svn+ssh://lumo.ucsd.edu/projects/p1/svnroot/pdollar/toolbox@12374 52fe0c90-79fe-0310-8a18-a0b98ad248f8

+++ images/private/ktComputeW_c.c
@@ -44,7 +44,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
   B = (uchar*) mxGetData(prhs[0]);
   q = mxGetPr(prhs[1]);
   p = mxGetPr(prhs[2]);
-  nBits = mxGetScalar(prhs[3]);
+  nBits = (int) mxGetScalar(prhs[3]);
   
   /* create output variable -- nBins = 2^nBits */
   dims[0]=n; dims[1]=1;
